requirements.txt
scipy
tldextract
opencv-python
pandas
numpy==1.23.0
tqdm
Pillow==8.4.0
pathlib
fvcore
pycocotools
scikit-learn
lxml
gdown
memory-profiler
psutil
torch
torchvision
